src/e/edit/TagReader.java: I've been having trouble on Ubuntu lately where Evergreen uses up all available file descriptors for pipes, leading to an inability to run "ctags" (java.io.Exception: error=24, Too many open files). This doesn't address the underlying problem, whatever it is, but it does mean I don't slowly fill /tmp with zero-length files while the situation persists.

+++ evergreen/trunk/src/e/edit/TagReader.java
@@ -18,12 +18,16 @@ public class TagReader {
         this.listener = tagListener;
         this.fileType = fileType;
         
+        File tagsFile = null;
         try {
-            File tagsFile = createTagsFileFor(file);
+            tagsFile = createTagsFileFor(file);
             readTagsFile(tagsFile);
-            tagsFile.delete();
         } catch (Exception ex) {
             listener.taggingFailed(ex);
+        } finally {
+            if (tagsFile != null) {
+                tagsFile.delete();
+            }
         }
     }
